All algorithms are rewritten, as are SQL queries, to add a filter parameter and to modify the input to vector layers instead of text parameters.
Some improvements are also made to the queries and scripts,
such as applying the script only to the selected features,
if such selection exists, for the script "Find Overlap In One Layer".

The translation was removed to ease maintenance and a script
was removed because other script does almost the same thing.

The documentation also was updated, and placed together
with the algorithm help string, in the case of English.